rewamp the UI
