--- app/src/main/java/com/epfl/drawyourpath/community/TournamentDetailFragment.kt
@@ -1,10 +1,0 @@
-package com.epfl.drawyourpath.community
-
-import androidx.fragment.app.Fragment
-import com.epfl.drawyourpath.R
-
-class TournamentDetailFragment : Fragment(R.layout.fragment_tournament_detail) {
-
-
-
-}
